+++ src/CSET/operators/plot.py
@@ -429,6 +429,173 @@ def _plot_and_save_vertical_line_series(
     plt.close(fig)
 
 
+def _plot_and_save_histogram_series(
+    cube: iris.cube.Cube,
+    filename: str,
+    title: str,
+    vmin: float,
+    vmax: float,
+    histtype: str = "step",
+    **kwargs,
+):
+    """Plot and save a histogram series.
+
+    Parameters
+    ----------
+    cube: Cube
+        2 dimensional Cube of the data to plot as histogram.
+        Plotting options are fixed:
+        density=True, histtype='step',stacked=True to ensure that
+        a probability density is plotted using matplotlib.pyplot.hist
+        to plot the probability density so that the area under
+        the histogram integrates to 1.
+        stacked is set to True so the sum of the histograms is
+        normalized to 1.
+        ax.autoscale is switched off and the ylim range
+        is preset as (0,1) to make figures comparable.
+    filename: str
+        Filename of the plot to write.
+    title: str
+        Plot title.
+    vmin: float
+        minimum for colourbar
+    vmax: float
+        maximum for colourbar
+    histtype: str
+        The type of histogram to plot. Options are "step" for a line
+        histogram or "barstacked", "stepfilled". "Step" is the default option,
+        but can be changed in the rose-suite.conf configuration.
+    """
+    fig = plt.figure(figsize=(8, 8), facecolor="w", edgecolor="k")
+    # Reshape cube data into a single array to allow for a single histogram.
+    # Otherwise we plot xdim histograms stacked.
+    cube_data_1d = (cube.data).flatten()
+    plt.hist(cube_data_1d, density=True, histtype=histtype, stacked=True)
+    ax = plt.gca()
+
+    # Add some labels and tweak the style.
+    ax.set(
+        title=title,
+        xlabel=f"{cube.name()} / {cube.units}",
+        ylabel="normalised probability density",
+        ylim=(0, 1),
+        xlim=(vmin, vmax),
+    )
+
+    # Save plot.
+    fig.savefig(filename, bbox_inches="tight", dpi=150)
+    logging.info("Saved line plot to %s", filename)
+    plt.close(fig)
+
+
+def _plot_and_save_postage_stamp_histogram_series(
+    cube: iris.cube.Cube,
+    filename: str,
+    title: str,
+    stamp_coordinate: str,
+    vmin: float,
+    vmax: float,
+    histtype: str,
+    **kwargs,
+):
+    """Plot and save postage (ensemble members) stamps for a histogram series.
+
+    Parameters
+    ----------
+    cube: Cube
+        2 dimensional Cube of the data to plot as histogram.
+        Plotting options are fixed:
+        density=True, histtype='bar', stacked=True to ensure that
+        a probability density is plotted using matplotlib.pyplot.hist
+        to plot the probability density so that the area under
+        the histogram integrates to 1.
+        stacked is set to True so the sum of the histograms is
+        normalized to 1.
+        ax.autoscale is switched off and the ylim range
+        is preset as (0,1) to make figures comparable.
+    filename: str
+        Filename of the plot to write.
+    title: str
+        Plot title.
+    stamp_coordinate: str
+        Coordinate that becomes different plots.
+    vmin: float
+        minimum for pdf x-axis
+    vmax: float
+        maximum for pdf x-axis
+    histtype: str
+        The type of histogram to plot. Options are "step" for a line
+        histogram or "barstacked", "stepfilled". "Step" is the default option,
+        but can be changed in the rose-suite.conf configuration.
+
+    """
+    # Use the smallest square grid that will fit the members.
+    grid_size = int(math.ceil(math.sqrt(len(cube.coord(stamp_coordinate).points))))
+
+    fig = plt.figure(figsize=(10, 10), facecolor="w", edgecolor="k")
+    # Make a subplot for each member.
+    for member, subplot in zip(
+        cube.slices_over(stamp_coordinate), range(1, grid_size**2 + 1), strict=False
+    ):
+        # Implicit interface is much easier here, due to needing to have the
+        # cartopy GeoAxes generated.
+        plt.subplot(grid_size, grid_size, subplot)
+        # Reshape cube data into a single array to allow for a single histogram.
+        # Otherwise we plot xdim histograms stacked.
+        member_data_1d = (member.data).flatten()
+        plt.hist(member_data_1d, density=True, histtype=histtype, stacked=True)
+        ax = plt.gca()
+        ax.set_title(f"Member #{member.coord(stamp_coordinate).points[0]}")
+        ax.set_ylim(0, 1)
+        ax.set_xlim(vmin, vmax)
+
+    # Overall figure title.
+    fig.suptitle(title)
+
+    fig.savefig(filename, bbox_inches="tight", dpi=150)
+    logging.info("Saved histogram postage stamp plot to %s", filename)
+    plt.close(fig)
+
+
+def _plot_and_save_postage_stamps_in_single_plot_histogram_series(
+    cube: iris.cube.Cube,
+    filename: str,
+    title: str,
+    stamp_coordinate: str,
+    vmin: float,
+    vmax: float,
+    histtype: str = "step",
+    **kwargs,
+):
+    fig, ax = plt.subplots(figsize=(10, 10), facecolor="w", edgecolor="k")
+    ax.set_title(title)
+    ax.set_xlim(vmin, vmax)
+    ax.set_ylim(0, 1)
+    ax.set_xlabel(f"{cube.name()} / {cube.units}")
+    ax.set_ylabel("normalised probability density")
+    # Loop over all slices along the stamp_coordinate
+    for member in cube.slices_over(stamp_coordinate):
+        # Flatten the member data to 1D
+        member_data_1d = member.data.flatten()
+        # Plot the histogram using plt.hist
+        plt.hist(
+            member_data_1d,
+            density=True,
+            histtype=histtype,
+            stacked=True,
+            label=f"Member #{member.coord(stamp_coordinate).points[0]}",
+        )
+
+    # Add a legend
+    ax.legend()
+
+    # Save the figure to a file
+    plt.savefig(filename)
+
+    # Close the figure
+    plt.close(fig)
+
+
 ####################
 # Public functions #
 ####################
@@ -656,10 +823,131 @@ def plot_vertical_line_series(
     if filename is None:
         filename = slugify(recipe_title)
 
-    # Make vertical line plot
-    plotting_func = _plot_and_save_vertical_line_series
-
     # set the lower and upper limit for the x-axis to ensure all plots
+    # have same range. This needs to read the whole cube over the range of
+    # the sequence and if applicable postage stamp coordinate.
+    # This only works if the plotting is done in the collate section of a
+    # recipe and not in the parallel section of a recipe.
+    vmin = np.floor(cube.data.min())
+    vmax = np.ceil(cube.data.max())
+
+    # Create a plot for each value of the sequence coordinate.
+    plot_index = []
+    for cube_slice in cube.slices_over(sequence_coordinate):
+        # Use sequence value so multiple sequences can merge.
+        seq_coord = cube_slice.coord(sequence_coordinate)
+        sequence_value = seq_coord.points[0]
+        plot_filename = f"{filename.rsplit('.', 1)[0]}_{sequence_value}.png"
+        # Format the coordinate value in a unit appropriate way.
+        title = f"{recipe_title} | {seq_coord.units.title(sequence_value)}"
+        # Do the actual plotting.
+        _plot_and_save_vertical_line_series(
+            cube_slice,
+            coord,
+            plot_filename,
+            series_coordinate,
+            title=title,
+            vmin=vmin,
+            vmax=vmax,
+        )
+        plot_index.append(plot_filename)
+
+    # Add list of plots to plot metadata.
+    complete_plot_index = _append_to_plot_index(plot_index)
+
+    # Make a page to display the plots.
+    _make_plot_html_page(complete_plot_index)
+
+    return cube
+
+
+def plot_histogram_series(
+    cube: iris.cube.Cube,
+    filename: str = None,
+    sequence_coordinate: str = "time",
+    stamp_coordinate: str = "realization",
+    single_plot: bool = False,
+    histtype: str = "step",
+    **kwargs,
+) -> iris.cube.Cube:
+    """Plot a histogram plot for each vertical level provided.
+
+    A histogram plot can be plotted, but if the sequence_coordinate (i.e. time)
+    is present then a sequence of plots will be produced using the time slider
+    functionality to scroll through histograms against time.
+    If a stamp_coordinate is present then postage stamp plots will be produced.
+    If stamp_coordinate and single_plot is True,
+    all postage stamp plots will be plotted in a single plot instead of separate postage stamp plots.
+
+    Parameters
+    ----------
+    cube: Cube
+        Iris cube of the data to plot. It should have a single dimension other
+        than the stamp coordinate.
+    filename: str, optional
+        Name of the plot to write, used as a prefix for plot sequences. Defaults
+        to the recipe name.
+    sequence_coordinate: str, optional
+        Coordinate about which to make a plot sequence. Defaults to ``"time"``.
+        This coordinate must exist in the cube and will be used for the time slider.
+    stamp_coordinate: str, optional
+        Coordinate about which to plot postage stamp plots. Defaults to
+        ``"realization"``.
+    single_plot: bool, optional
+        If True, all postage stamp plots will be plotted in a single plot.
+        If False, each postage stamp plot will be plotted separately.
+        Is only valid if stamp_coordinate exists and has more than a single point.
+    histtype: str, optional
+        The type of histogram to plot. Options are "step" for a line
+        histogram or "barstacked", "stepfilled". "Step" is the default option,
+        but can be changed in the rose-suite.conf configuration.
+
+    Returns
+    -------
+    Cube
+        The original cube (so further operations can be applied).
+
+    Raises
+    ------
+    ValueError
+        If the cube doesn't have the right dimensions.
+    TypeError
+        If the cube isn't a single cube.
+    """
+    recipe_title = get_recipe_metadata().get("title", "Untitled")
+
+    # Ensure we have a name for the plot file.
+    if filename is None:
+        filename = slugify(recipe_title)
+
+    # Ensure we've got a single cube.
+    cube = _check_single_cube(cube)
+
+    # Internal plotting function.
+    plotting_func = _plot_and_save_histogram_series
+
+    # Make postage stamp plots if stamp_coordinate exists and has more than a
+    # single point. If single_plot is True, all postage stamp plots will be
+    # plotted in a single plot instead of separate postage stamp plots.
+    try:
+        if cube.coord(stamp_coordinate).shape[0] > 1:
+            if single_plot:
+                plotting_func = (
+                    _plot_and_save_postage_stamps_in_single_plot_histogram_series
+                )
+            else:
+                plotting_func = _plot_and_save_postage_stamp_histogram_series
+    except iris.exceptions.CoordinateNotFoundError:
+        pass
+
+    # If several histograms are plotted with time as sequence_coordinate for the
+    # time slider option.
+    try:
+        cube.coord(sequence_coordinate)
+    except iris.exceptions.CoordinateNotFoundError as err:
+        raise ValueError(f"Cube must have a {sequence_coordinate} coordinate.") from err
+
+    # Set the lower and upper limit for the colorbar to ensure all plots
     # have same range. This needs to read the whole cube over the range of
     # the sequence and if applicable postage stamp coordinate.
     # This only works if the plotting is done in the collate section of a
@@ -673,18 +961,18 @@ def plot_vertical_line_series(
         # Use sequence value so multiple sequences can merge.
         sequence_value = cube_slice.coord(sequence_coordinate).points[0]
         plot_filename = f"{filename.rsplit('.', 1)[0]}_{sequence_value}.png"
-        coord = cube_slice.coord(series_coordinate)
+        coord = cube_slice.coord(sequence_coordinate)
         # Format the coordinate value in a unit appropriate way.
         title = f"{recipe_title} | {coord.units.title(coord.points[0])}"
         # Do the actual plotting.
         plotting_func(
             cube_slice,
-            coord,
             plot_filename,
-            series_coordinate,
+            stamp_coordinate=stamp_coordinate,
             title=title,
             vmin=vmin,
             vmax=vmax,
+            histtype=histtype,
         )
         plot_index.append(plot_filename)